refactor(playwright): use dequeUtitilies sub module for JSON mapping classes
