--- playwright/src/main/java/com/deque/html/axecore/utilities/axeresults/AxeRuntimeException.java
@@ -1,9 +1,0 @@
-package com.deque.html.axecore.utilities.axeresults;
-
-public class AxeRuntimeException extends RuntimeException {
-  private static final long serialVersionUID = -123456789087654L;
-
-  public AxeRuntimeException(Exception cause) {
-    super("Error when running axe", cause);
-  }
-}
