--- playwright/src/main/java/com/deque/html/axecore/utilities/axeresults/FrameContext.java
@@ -1,22 +1,0 @@
-package com.deque.html.axecore.utilities.axeresults;
-
-public class FrameContext {
-  private Object frameSelector;
-  private Object frameContext;
-
-  public Object getFrameSelector() {
-    return frameSelector;
-  }
-
-  public void setFrameSelector(final Object selector) {
-    this.frameSelector = selector;
-  }
-
-  public Object getFrameContext() {
-    return frameContext;
-  }
-
-  public void setFrameContext(final Object context) {
-    this.frameContext = context;
-  }
-}
