--- playwright/src/main/java/com/deque/html/axecore/utilities/axeresults/Node.java
@@ -1,24 +1,0 @@
-package com.deque.html.axecore.utilities.axeresults;
-
-public class Node {
-  private String html;
-  private Object target;
-
-  public Node() {}
-
-  public String getHtml() {
-    return this.html;
-  }
-
-  public void setHtml(String html) {
-    this.html = html;
-  }
-
-  public Object getTarget() {
-    return this.target;
-  }
-
-  public void setTarget(Object target) {
-    this.target = target;
-  }
-}
