--- playwright/src/main/java/com/deque/html/axecore/utilities/axeresults/TestEngine.java
@@ -1,20 +1,0 @@
-package com.deque.html.axecore.utilities.axeresults;
-
-public class TestEngine {
-  private String name;
-  private String version;
-
-  public TestEngine() {}
-
-  public String getName() {
-    return this.name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getVersion() {
-    return this.version;
-  }
-}
