--- playwright/src/main/java/com/deque/html/axecore/utilities/axeresults/TestRunner.java
@@ -1,11 +1,0 @@
-package com.deque.html.axecore.utilities.axeresults;
-
-public class TestRunner {
-  private String name;
-
-  public TestRunner() {}
-
-  public String getName() {
-    return this.name;
-  }
-}
